Revision summary: -Cleaned up parts of the code -Removed some unused code -Removed couple of warnings -Optimized some parts of the code +Added new logging system
Revision notes:
-The logging system is not yet
fully embedded.
Most of the critical parts do use
the new logging system for additional
speed when logging is disabled(Default state).

-(Temporary)To enable back full logging,
open /nullDc/log/log.cpp

Replace :
mMode(LOG_MODE_SILENT)

With :
mMode(LOG_MODE_DEFAULT)

The list of all available modes can be found in log.hpp.

Additional notes to fellow programmers:

Please try to use the new logging system.
Look at the interface how to use it.Its as simple
as replacing 'printf' with 'logWrite' macro.

To switch modes use logSetMode(mode) or logSetModeEx(mode,arg0).

ie - if you wish to redirect the log to a file , do this :
logSetModeEx(LOG_MODE_STREAM,filepath);

+++ nulldc/nullDC/dc/asic/asic.cpp
@@ -75,7 +75,7 @@ void RaiseAsicErr(HollyInterruptID inter)
 
 void fastcall asic_RaiseInterrupt(HollyInterruptID inter)
 {
-	u8 m=inter>>8;
+	u8 m=(u8)(inter>>8);
 	switch(m)
 	{
 	case 0:
